Added JSON handlers for solid_frame and titled_frame
