+++ include/munin/titled_frame.hpp
@@ -46,6 +46,13 @@ public :
         std::shared_ptr<component> const &associated_component);
 
 private :
+    //* =====================================================================
+    /// \brief Called by to_json().  Derived classes must override this
+    /// function in order to add additional data about their implementation
+    /// in a custom manner.
+    //* =====================================================================
+    nlohmann::json do_to_json() const override;
+
     struct impl;
     std::shared_ptr<impl> pimpl_;
 };
