+++ test/src/solid_frame/solid_frame_test.hpp
@@ -1,0 +1,37 @@
+#pragma once
+#include "mock/render_surface_capabilities.hpp"
+#include <munin/solid_frame.hpp>
+#include <gtest/gtest.h>
+#include <gmock/gmock.h>
+
+class a_solid_frame_with_no_unicode_support : public testing::Test
+{
+protected :
+    void SetUp() override
+    {
+        using testing::Return;
+
+        ON_CALL(surface_capabilities_, supports_unicode())
+            .WillByDefault(Return(false));
+    }
+
+    testing::NiceMock<mock_render_surface_capabilities> surface_capabilities_;
+    munin::solid_frame frame_;
+};
+
+using a_solid_frame = a_solid_frame_with_no_unicode_support;
+
+class a_solid_frame_with_unicode_support : public testing::Test
+{
+protected :
+    void SetUp() override
+    {
+        using testing::Return;
+
+        ON_CALL(surface_capabilities_, supports_unicode())
+            .WillByDefault(Return(true));
+    }
+
+    testing::NiceMock<mock_render_surface_capabilities> surface_capabilities_;
+    munin::solid_frame frame_;
+};
